Ending ai_modernization work
We have reached near the end of the line for the modernization work and
are now at a point where major changes in the software are needed to
refactor the software. Currently the software will compile on a 64 bit
linux system and has extensive documentation provided by claude for all
routines and other items. At this point we need to begin reintegrating
this code into a new 'main' branch and then start on future work with an
idea of producing a 4.20 release which would refactor the code to be
testable and then rework towards the 5.0 code.

Signed-off-by: Stephen Smoogen <ssmoogen@redhat.com>